Changed quaternion class /include/cyclone/core.h
It is rather unusual to have a perfect equality when comparing floating-point numbers. Thus it is necessary to compare with a margin, it being the machine epsilon. 

It is not needed to check against the abs of d, as it is made from (hopefully) positive numbers. If, by any chance, d<0, will be within machine error, and also caught by the current condition.